rename package cloudsqlutil to util
